nykaa-jit/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.34.0
python-dotenv==1.0.0